+++ cmArt.WebJaguar.Connector/Data/Product_Price.cs
@@ -7,13 +7,13 @@ namespace cmArt.WebJaguar.Connector
     // Root myDeserializedClass = JsonConvert.DeserializeObject<Root>(myJsonResponse); 
     public class Product_Price
     {
-        public int amt { get; set; }
-        public int discountAmt { get; set; }
+        public double amt { get; set; }
+        public double discountAmt { get; set; }
         public int qtyFrom { get; set; }
         public int qtyTo { get; set; }
-        public int caseAmt { get; set; }
-        public int caseDiscountAmt { get; set; }
-        public int cost { get; set; }
+        public double caseAmt { get; set; }
+        public decimal caseDiscountAmt { get; set; }
+        public decimal cost { get; set; }
     }
